# palikka-api/conf/docs/users/api-docs.yaml
openapi: 3.0.1
info:
  contact:
    email: janne.tahkola@gmail.com
    name: Janne Tahkola
    url: https://github.com/jannetahkola
  description: User & access management
  title: Users API
  version: 1.0.0
servers:
- description: Generated server url
  url: http://localhost:8080/users-api
security:
- Authorization: []
paths:
...
  /privileges:
    get:
      description: Results may be filtered depending on the user's authorities
      operationId: getPrivileges
      parameters:
      - description: |
          Results are filtered into those containing the search text either in the privilege's domain or name
          - Case insensitive
          - Between 1-6 letters and/or numbers
        example: ban
        in: query
        name: search
        required: false
        schema:
          type: string
      - description: Bearer access token
        example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzUxMiIsInB0...
        in: header
        name: Authorization
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelPrivilegeModel'
          description: OK
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Forbidden
        "405":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Method Not Allowed
        "406":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Acceptable
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Internal Server Error
      summary: Get all privileges
      tags:
      - Privileges
  /roles:
    get:
      description: Results may be filtered depending on the user's authorities
      operationId: getRoles
      parameters:
      - description: Bearer access token
        example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzUxMiIsInB0...
        in: header
        name: Authorization
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelRoleModel'
          description: OK
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Forbidden
        "405":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Method Not Allowed
        "406":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Acceptable
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Internal Server Error
      summary: Get all roles
      tags:
      - Roles
...
  /roles/{role_id}/privileges:
    get:
      operationId: getRolePrivileges
      parameters:
      - in: path
        name: role_id
        required: true
        schema:
          type: integer
          format: int32
      - description: Bearer access token
        example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzUxMiIsInB0...
        in: header
        name: Authorization
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelPrivilegeModel'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/CollectionModelPrivilegeModel'
          description: OK
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Forbidden
        "405":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Method Not Allowed
        "406":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Acceptable
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Internal Server Error
      summary: Get all privileges associated with a role
      tags:
      - Role privileges
    post:
      operationId: postRolePrivileges
      parameters:
      - description: Identifier of the role
        example: 1
        in: path
        name: role_id
        required: true
        schema:
          type: integer
          format: int32
      - description: Bearer access token
        example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzUxMiIsInB0...
        in: header
        name: Authorization
        required: true
        schema:
          type: string
      - description: JSON content type
        example: application/json
        in: header
        name: Content-Type
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RolePrivilegePostModel'
        required: true
      responses:
        "201":
          content:
            application/hal+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PrivilegeModel'
          description: Created
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Bad Request
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Found
        "405":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Method Not Allowed
        "406":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Acceptable
        "415":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Unsupported Media Type
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Internal Server Error
      summary: Create a new privilege association for a role
      tags:
      - Role privileges
  /roles/{role_id}/privileges/{privilege_id}:
    delete:
      operationId: deleteRolePrivileges
      parameters:
      - description: Identifier of the role
        example: 1
        in: path
        name: role_id
        required: true
        schema:
          type: integer
          format: int32
      - description: Identifier of the privilege
        example: 1
        in: path
        name: privilege_id
        required: true
        schema:
          type: integer
          format: int32
      - description: Bearer access token
        example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzUxMiIsInB0...
        in: header
        name: Authorization
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Bad Request
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Found
        "405":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Method Not Allowed
        "406":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Acceptable
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Internal Server Error
      summary: Delete a privilege association from a role
      tags:
      - Role privileges
    get:
      operationId: getRolePrivilege
      parameters:
      - description: Identifier of the role
        example: 1
        in: path
        name: role_id
        required: true
        schema:
          type: integer
          format: int32
      - description: Identifier of the privilege
        example: 1
        in: path
        name: privilege_id
        required: true
        schema:
          type: integer
          format: int32
      - description: Bearer access token
        example: Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzUxMiIsInB0...
        in: header
        name: Authorization
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/PrivilegeModel'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/PrivilegeModel'
          description: OK
        "403":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Forbidden
        "404":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Found
        "405":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Method Not Allowed
        "406":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Not Acceptable
        "500":
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
          description: Internal Server Error
      summary: Get a privilege associated with a role
      tags:
      - Role privileges
...
components:
  schemas:
    CollectionModelPrivilegeModel:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            privileges:
              type: array
              items:
                $ref: '#/components/schemas/PrivilegeModel'
        _links:
          $ref: '#/components/schemas/Links'
    CollectionModelRoleModel:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            roles:
              type: array
              items:
                $ref: '#/components/schemas/RoleModel'
        _links:
          $ref: '#/components/schemas/Links'
...
    Link:
      type: object
      properties:
        deprecation:
          type: string
        href:
          type: string
        hreflang:
          type: string
        name:
          type: string
        profile:
          type: string
        templated:
          type: boolean
        title:
          type: string
        type:
          type: string
    Links:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/Link'
...
    PrivilegeModel:
      type: object
      description: Authorizes users with associated roles to perform various actions
        within APIs
      properties:
        _links:
          $ref: '#/components/schemas/Links'
        domain:
          type: string
          description: Domain of the privilege. References an action inside an API
            this privilege is applied within
          example: COMMAND
        domain_description:
          type: string
          description: An optional text to describe the privilege in its domain
          example: Adds IP address to banlist.
        id:
          type: integer
          format: int32
          description: Identifier of the privilege
          example: 1234
        name:
          type: string
          description: Name of the privilege. Provides more fine grained access control
            within a domain
          example: weather
      required:
      - domain
      - id
      - name
    ProblemDetail:
      type: object
      properties:
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties:
            type: object
        status:
          type: integer
          format: int32
        title:
          type: string
        type:
          type: string
          format: uri
...
    RoleModel:
      type: object
      description: Authorizes associated users to access different APIs
      properties:
        _links:
          $ref: '#/components/schemas/Links'
        description:
          type: string
          description: Description of the role
          example: Access to limited functionality
        id:
          type: integer
          format: int32
          description: Identifier of the role
          example: 1234
        name:
          type: string
          description: Unique name of the role
          example: ROLE_USER
        privileges:
          type: array
          description: Privileges associated with the role
          items:
            $ref: '#/components/schemas/PrivilegeModel'
          readOnly: true
      required:
      - id
      - name
    RolePrivilegePostModel:
      type: object
      description: Request object containing the identifier of the privilege that
        should be associated with the role
      properties:
        privilege_id:
          type: integer
          format: int32
          description: Identifier of the privilege to associate with the role
          example: 1
      required:
      - privilege_id
...
  securitySchemes:
    Authorization:
      bearerFormat: JWT
      description: JWT access token
      in: header
      scheme: Bearer
      type: http